backend/requirements.txt
asgiref==3.8.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dataclasses==0.6
distlib==0.3.8
Django==5.0.7
django-admin==2.0.2
django-cors-headers==4.4.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.2
djongo==1.2.31
dnspython==2.6.1
excel-base==1.0.4
filelock==3.15.4
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
isoweek==1.3.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
platformdirs==4.2.2
psycopg2-binary==2.9.9
pygame==2.5.2
pymongo==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
screen==1.0.1
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.5.1
TimeConvert==3.0.13
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
unflatten==0.1.1
virtualenv==20.26.3
Werkzeug==3.0.1
xlwt==1.3.0
